Timing_Analyzer/requirements.txt
pyserial==3.5
pandas==2.0.3
openpyxl==3.1.2
